# api/openapi.yaml
openapi: "3.0.3"
info:
  title: TurnScore API
  version: "1.0.0"
  description: |
    Tabletop Tournament Table Rating API.
    Organizers manage tournaments and tables; participants rate them via QR code.
servers:
  - url: /api/v1

# ---------------------------------------------------------------------------
# Security Schemes
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # -------------------------------------------------------------------------
  # Reusable Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "not_found"
        message:
          type: string
          example: "Tournament not found"

    # --- User / Auth ---
...
    AuthTokens:
      type: object
      required: [access_token, refresh_token, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          description: Access token lifetime in seconds
          example: 900
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string

    # --- Tournament ---
...
  responses:
...
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /auth/refresh:
    post:
      summary: Refresh access token
      operationId: refreshToken
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
        "401":
          $ref: "#/components/responses/Unauthorized"
...